Killing static field inspector, and instead using Mutability inspector. The analyzer now assess immutability of statics and reports unsafety there as an error.
Mutability is considered at the field/property level only; method purity is not considered (at least yet). Mutability is also only considered as it relates to the class externally (fields or properties with private accessors or private setters are considered immutable).
A type is considered immutable iff (1) all of its fields/properties are immutable and (2) all of the types of said fields/properties are also immutable.