Making all dependencies in the UnsafeStaticCounter a dev dependency.
AppVeyor calls `nuget pack` in the csproj's context, which asigns all dependencies in the `packages.config` file as NuGet dependencies in the final NuGet package. This is not necessary, because the tool bundles all dependencies with it. Instead of overwriting AppVeyor's NuGet packing behaviour (and having to pass in all variables to `nuget pack` ourselves), we just mark those dependencies as development dependencies, to prevent them from being required for the final package.