Simplify type symbol comparison
The helper method IsPropertyOfType wasn't helpful; it was more verbose
than just using .Equals. It was necessary due to a poor choice for the
order of .Equals (because one (but only one) argument could be null.)
Switching that order and inlining the function made it more readable. It
also removed redundant lookups in the semantic model.

The other helper is still useful because it takes care of handling the
semantic model and is used within a linq expression.

Equals was used because it is most obviously value equality. Reading the
Roslyn source code indicates that == is overrided to do the same thing
but I'm not sure the readability improvement is worth the alarm it could
cause when reading the code.