Emit diagnostics for unnecessary and conflicting static annotations
This required refactoring InspectType (renamed to InspectFieldOrProperty
because that is more accurate) and the AnalyzeField/Property methods.

In the new design most branches out of analysis are via yield breaks in
GenerateDiagnostics. This allows InspectFieldOrProperty to implement the
"if there are no diagnostics but there is an annotation, emit a
diagnostic" logic. Bailing out earlier (e.g. in AnalyzeField) would
break this. I've added some more comments, particularly around any
bailing-out and some overall design notes for future readers.

In the near future I plan to fix the way attributes are compared to use
types from semantic analysis similar to the RpcAnalyzer.