Add spec tests prototype
This form of testing for the analyzers would be awesome I think.

- We get VS syntax highlighting for the files
- There is minimal boilerplate to add new test cases
- The assertions are beside where the diagnostics should be emitted
- We can delineate where the diagnostic appears so we can assert where
  the squiggle should go without manually having to type line and column
  numbers.
- We could generate pretty HTML from this: remove the comments, syntax
  highlight, add red-squiggles/error messages.

The biggest win though, in my opinion, is for TDD like development: an
analyzer author can write a spec file full of evil code that should have
diagnostics and then write the analyzer with the assistance of the
VS syntax tree visualizer on the spec file.

This is just a prototype and its got some sketchy code in there. It's
only asserting that the number of diagnostics matches the number of
non-empty comments.