Fix the logic for analyzing properties
The helper method for PropertySyntax was changed from
IsPropertyGetterImplemented to IsAutoProperty. IsAutoProperty checks
strictly more things than IsPropertyGetterImplemented but its also the
reverse bool (i.e. IsPropertyGetterImplemented => !IsAutoProperty.)

Properties should only be treated like fields if they are
auto-properties (which includes "readonly properties" which are similar
to readonly fields when it comes to mutability.) This is reflected in
the logic of the analyzer by bailing out of analysis when there is an
auto-property.

There are now two bools due to the logic flip. It made sense to pass
false for IsPropertyGetterImplemented for fields because... they aren't
properties. We could send true for IsAutoProperty for fields (since they
behave similarly to auto properties) but we usually think of properties
being lowered to fields so being explicit reads better. A three-way enum
could be used (Field, NonAutoProperty, AutoProperty) but the booleans
are perfectly readable.