Hack the relevant analyzers to import all metadata.
We do this because analyzers can behave differently depending on what context they're running in (see: https://github.com/dotnet/roslyn/issues/20373).

This should solve the problem, because as far as I was able to dig into the code, member enumeration is completely lazy and checks that option from `Compilation.Option` all the time. This works in `msbuild`, but Visual Studio seems to override the behaviour. Still working on that...

We also just throw completely if the options are not set correctly, as we can't be safe.