Adding support for assembly-declared known immutables.
The idea behind this is that we can change the list of known immutable types without release a new version of the analyzer. All projects in the LMS will reference a single file in LP that has them all defined.

I'll be following a similar idea for container types, but that's non-trivial because of the field/props we use for the type arguments of those containers.