US87654: Make [NotNull] parameter caching more meaningful
The value is never set to `null`, but instead is always set to an immutable hash set based on the not null parameters from an earlier run.
This means that comparing it to `null` is meaningless, since if the value is ever retrieved it will not be `null`, and instead the count should be checked.

I'm also stopping it from creating a new hash set every time, even if it's already been cached.