Add ClassGoal and StructGoal reduction rules
This is the first PR in a redesign of the mutability inspector I've been
working on. It's an arbitrary but representative slice of the internals.
My plan is to merge this in "on the side" and not wire it into the
analyzers until it is complete.

Internally we operate on goals by applying rules to reduce them to
subgoals. Example goals include Class, Struct, ConcreteType, Type,
Property, etc. A rule is a static function with a signature like

`static IEnumerable<Goal> Apply( SemanticModel model, Goal goal );`

(The argument may be a specific subtype of goal.)

If a rule reduces a goal to no subgoals then it is definitevely not
mutable. If it returns the goal itself back that means it couldn't be
reduced. Otherwise rules won't emit a goal as a subgoal of itself.

If a rule can't reduce a goal there may be another rule that can. An
example of this is that we have a list of Types and a list of
ConcreteTypes which are known to be safe.

Later, there will be logic to break infinite loops (by maintaining a
concurrent hash set of visited types) and caching of results for types.
Towards the end I'll insturment the code that runs the rules (the
mutability inspector) with calls to a logging interface. The purpose of
logging will be to allow outside users of the mutability inspector to
see which rules were applied (with their inputs and outputs) so that we
can build a debugging tools for the analyzer.