Minor cleanup of InspectProperty and InspectField.
GetDeclarationSyntax will also be used for fields once we start looking
at initializers. It's pretty picky that there is exactly one
declaration. Previously, properties were a bit more flexible about that
and would continue to analyze the type of the property even if it wasn't
an auto-property. I don't think that's correct even if it were possible
that there were 0 or more than 1 decl for the property (I don't think it is.)