Commute the check for interfaces up to InspectTypeRecursive
This change should be a no-op: interfaces don't recurse so the typestack
check was unnecessary. As a small side benefit these types never appear
in type stack now.

The InspectClassStructOrInterface method was weird. When I refactored
that stuff iirc I meant for it to be InspectClassOrStruct but fucked up.

There is only one more commutation I'm going to do; I did this in a
bunch of "easy" pieces to avoid one hard-to-think-about move.