Commute check for unsealed classes
This is safe because its only skipping over the type stack check but
there is no recursion in this check.

Also, remove inspection flags from InspectClassOrStruct.

It might seem a bit weird to have class-specific logic outside of
InspectClassOrStruct. What I'm going for (ultimately, not today) is to
generalize this into a type vs. concrete type subgoal that's independent
of scanning the class decl (members, base etc.) It's a constraint that
applies to more than classes, but different sorts of types satisfy it in
different ways (e.g. interfaces via [Immutable], or in the future
possibly via being internal/private and having only immutable impls.)