Add caching to MutabilityInspector
This fixes the worst perf problems of the analyzer that have recently
plagued the build (e.g. the worst case perf is no longer exponential.)

There is still some work to do: the analyzers should share a cache and
if we've validated a type is immutable then the concrete type (i.e.
analyzing with AllowUnsealed) should be a cache hit. This gets the build
back to normal ASAP.