Fix symbol equality comparisons
Prior to MSBuild 16.5, these symbol reference equality comparisons worked. This has been confirmed by analyzing known-broken code on Visual Studio 2019 16.4. With the update to MSBuild 16.5+, these started silently failing instead.

The 3.6.0 release of the CodeAnalysis packages includes an analyzer that catches these invalid equality checks, which was used to identify all of the broken cases.

This also updates the test projects to use the 3.6.0 release of the CodeAnalysis packages. These packages align with the Roslyn version provided with [Visual Studio 2019 and .NET Core 3.1][0].

[0]: https://github.com/dotnet/roslyn/blob/master/docs/wiki/NuGet-packages.md#versioning