remove ImmutableGenericArgumentAnalyzer
This is largely covered via the work that ImmutabilityAnalyzer does.
IGAA would also check for implied immutabilty requirements through type
constraints, but only for constraints on the parameter itself. For
example, the following would raise a diagnostic, requiring T be marked
[Immutable]:

    interface IFoo<[Immutable] S> {}

    class Foo<T> where T : IFoo<T> {}

While this is interesting, it is only useful for recursive definitions,
which are very rare. A more useful version might cover the following,
requiring U be marked [Immutable]:

    interface IFoo<[Immutable] S> {}

    class Foo<T, U> where T : IFoo<U> {}

Given this is not captured, it's utility is minimal and can be ignored.
Such a diagnostic is a helper, but does not add safety.