Default to .NET Standard 2.0 for all projects
Moving this to the default because it makes sense: "annotations"-style
DLLs we ship should probably target that for simplicity (the existing
Annotations DLL also multi-targets net20 which we can probably drop?),
and the analyzers have to target netstandard2.0.

Our test projects will need to override this; currently we run our tests
in net48 for some reason, but we should target net7/8/etc. I think.

Overall I think this will keep the csproj's simpler, especially once I
split the packages.