RequireNamedArgumentsAnalyzer: add support for exemptions
ExemptSymbolsBuilder is a helper for building sets of exempt symbols
without bespoke logic per-analyzer.

The AdditionalFile format is similar to Roslyn's BannedApiAnalyzer:
https://github.com/dotnet/roslyn-analyzers/blob/3211f48253bc18560156d90dc5e710d35f7d03fa/src/Microsoft.CodeAnalysis.BannedApiAnalyzers/BannedApiAnalyzers.Help.md

In this case the optional [;Description Text] is excluded from the
format, as exemptions do not lead to user-facing messages.

Comments are supported for developers to leave notes justifying the exemption.